requirements.txt
requests
ipwhois
dfir-unfurl
gpt4all
pywin32; sys_platform == 'win32'
yara-python
pyyaml
extract_msg